+++ app/models/survey.rb
@@ -10,14 +10,13 @@ class Survey < ApplicationRecord
   has_rich_text :intro_en
   has_rich_text :intro_cy
 
-  def self.trigger_for(user, message_count:, last_message: false)
+  def self.trigger_for(user, message_count:)
     already_sent_ids = user.survey_sends.select(:survey_id)
 
     surveys = where.not(id: already_sent_ids)
 
     matching = surveys.select do |survey|
-      (survey.send_after_message_count == message_count) ||
-        (last_message && survey.send_on_last_message?)
+      survey.send_after_message_count == message_count
     end
 
     matching.each { |survey| SendSurveyJob.perform_later(user, survey) }