+++ db/migrate/20260423145235_add_programme_length_to_users.rb
@@ -1,0 +1,6 @@
+class AddProgrammeLengthToUsers < ActiveRecord::Migration[8.1]
+  def change
+    add_column :users, :programme_length, :integer
+    change_column_default :users, :programme_length, from: nil, to: 52
+  end
+end